GCC 7.x: Makefile: Fix device not booting
nathanchance: Filtered from the original commit (https://github.com/zanezam/boeffla-kernel-cm-n51x0/commit/228cc26b4b66ce3c9232eb00faaf95c3b1b3ae3c)

I tested all of the flags in the above commit and this is the only
one that prevented the phone from booting, with both -Os and -O3.

Without this, the phone will boot to the Google logo, hang for a
few seconds, then reboot, and repeat the process. I do not know why
this happens but this flag can be read about below. I don't think
we are missing much and it is a hell of a lot better than disabling
all of the new opts that make a new GCC version what it is.

https://gcc.gnu.org/onlinedocs/gcc-7.1.0/gcc/Optimize-Options.html#index-fstore-merging

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: spezi77 <spezi77@gmx.com>